requirements.txt
# For Google API Access (Gmail)
google-api-python-client==2.138.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# For data validation and modeling
pydantic==2.8.2